requirements.txt
cryptography==41.0.3      # For encryption, decryption, and key derivation
requests==2.31.0          # For HIBP API calls
pyperclip==1.8.2          # For clipboard operations
pyautogui==0.9.53         # For simulating keyboard and mouse actions
keyring==24.2.0           # For securely storing and retrieving keys
colorama==0.4.1           # For colored terminal output
tqdm==4.65.0              # For progress bars
argon2-cffi==21.3.0       # For Argon2 hashing
qrcode==7.4.2             # For generating QR codes
pyotp==2.8.0              # For generating TOTP codes
